# src/main/resources/swagger/swagger.yml
...
openapi: 3.0.0
info:
  title: Community REST API
  description: Community REST API
  version: "1.0"
...
servers:
  - url: http://localhost:8080/
  - url: https://localhost:8080/
################################################################################
#                               Components                                     #
################################################################################
components:
...
  parameters:
    OffsetParam:
      name: offset
      in: query
      description: Offset
      required: true
      schema:
        type: integer
        format: int32
        default: -1
    LimitParam:
      name: limit
      in: query
      description: Negative number means unlimited
      required: true
      schema:
        type: integer
        format: int32
        default: -1
...
  schemas:
...
    TopicType:
      type: string
      enum:
        - general
        - problem
        - system
        - eerpmGeneral
        - eerpmSummary
        - eerpmManual
        - eerpqGeneral
        - eerppGeneral
        - eerpmHighLevel
...
    TopicState:
      type: string
      enum:
        - unconcluded
        - concluded
        - briefConcluded
...
    TopicStatus:
      type: string
      enum:
        - open
        - delete
        - locked
...
    UserStatus:
      type: string
      enum:
        - active
        - inactive
      default: active
...
    ConclusionAlertMemberType:
      type: string
      enum:
        - group
        - user
...
    AlertRuleType:
      type: string
      enum:
        - general
        - high
    # 參考前端元件的key (前端不需多mapping)
...
    User:
      type: object
      properties:
        id:
          type: string
          default: ""
        name:
          type: string
          default: ""
        mail:
          type: string
          default: ""
        lock:
          type: boolean
          default: false
        roleId:
          type: integer
          format: int32
          default: 0
        totalCount:
          type: integer
          format: int32
          default: 0
        status:
          $ref: "#/components/schemas/UserStatus"
...
    IdNameDto:
      type: object
      properties:
        id:
          type: integer
          format: int32
        name:
          type: string
          default: ""
...
    TopicListDetailOfBookmark:
      type: object
      properties:
        id:
          type: integer
          format: int32
        title:
          type: string
        text:
          type: string
        type:
          $ref: "#/components/schemas/TopicType"
        state:
          $ref: "#/components/schemas/TopicState"
        status:
          $ref: "#/components/schemas/TopicStatus"
        createUser:
          $ref: "#/components/schemas/User"
        lastModifiedUser:
          $ref: "#/components/schemas/User"
        lastModifiedTime:
          type: integer
          format: int64
        belongForum:
          $ref: "#/components/schemas/IdNameDto"
        belongCommunity:
          $ref: "#/components/schemas/IdNameDto"
    TopicResultOfBookmark:
      type: object
      properties:
        result:
          type: array
          items:
            $ref: "#/components/schemas/TopicListDetailOfBookmark"
        numFound:
          type: integer
          format: int32
...
    LabelValueDto:
      type: object
      properties:
        label:
          type: string
          nullable: false
        value:
          type: object
...
    ConclusionAlertDetail:
      allOf:
        - $ref: "#/components/schemas/ConclusionAlertGroupDetail"
        - $ref: "#/components/schemas/ConclusionAlertRuleDetail"
        - $ref: "#/components/schemas/ConclusionAlertHighLevelRuleDetail"
        - type: object
          properties:
            ruleColumns:
              type: array
              items:
                $ref: "#/components/schemas/CellComponentDto"
            dropdown:
              type: object
              additionalProperties:
                type: array
                items:
                  $ref: "#/components/schemas/LabelValueDto"
              example:
                factory:
                  - label: DG2
                    value: 1
            rangeDay:
              type: object
              additionalProperties:
                $ref: "#/components/schemas/RangeDayDto"
              example:
                startDay:
                  isLimited: true
                  range:
                    - 0
                    - 30
    ConclusionAlertGroupDetail:
      type: object
      properties:
        groups:
          type: array
          items:
            $ref: "#/components/schemas/ConclusionAlertGroup"
        groupLastModifiedTime:
          type: integer
          format: int64
          default: 0
    ConclusionAlertGroup:
      type: object
      properties:
        key:
          type: string
          nullable: false
          default: ""
        label:
          type: string
          maxLength: 20
        members:
          type: array
          items:
            $ref: "#/components/schemas/ConclusionAlertMember"
    ConclusionAlertRuleDetail:
      type: object
      properties:
        rules:
          type: array
          items:
            $ref: "#/components/schemas/ConclusionAlertRule"
        ruleLastModifiedTime:
          type: integer
          format: int64
          default: 0
    ConclusionAlertHighLevelRuleDetail:
      type: object
      properties:
        highLevelRules:
          type: array
          items:
            $ref: "#/components/schemas/ConclusionAlertRule"
    ConclusionAlertRule:
      type: object
      properties:
        key:
          type: integer
          format: int32
          nullable: false
          default: 0
        startDay:
          type: integer
          format: int32
          minimum: 0
          maximum: 30
        endDay:
          type: integer
          format: int32
          minimum: -1
          maximum: 30
        members:
          type: array
          items:
            $ref: "#/components/schemas/ConclusionAlertMember"
        factory:
          $ref: "#/components/schemas/LabelValueDto"
    ConclusionAlertMember:
      type: object
      properties:
        key:
          type: string
        label:
          type: string
        type:
          $ref: "#/components/schemas/ConclusionAlertMemberType"
    CellComponentDto:
      type: object
      properties:
        cell:
          type: string
        component:
          type: string
    RangeDayDto:
      type: object
      properties:
        isLimited:
          type: boolean
        range:
          type: array
          items:
            type: integer
            format: int32
...
paths:
...
  /bookmark/{userId}/topic/bookmarks:
    get:
      tags:
        - Bookmark
      summary: List the bookmarks of topic
      operationId: getTopicListOfBookmark
      parameters:
        - name: userId
          in: path
          description: Id of the user
          required: true
          schema:
            type: string
        - $ref: "#/components/parameters/OffsetParam"
        - $ref: "#/components/parameters/LimitParam"
        - name: sort
          in: query
          description: 
            Sort
              * -bookmarkCreateTime - Sort by bookmark Create Time. Descending.
          required: true
          schema:
            type: string
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TopicResultOfBookmark"
        400:
          description: Invalid parameters
...
  /forum/{forumId}/alert:
    get:
      tags:
        - ConclusionAlert
      summary: Get alert rules in the forum
      operationId: getForumConclusionAlert
      parameters:
        - name: forumId
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConclusionAlertDetail"
        400:
          description: Invalid parameters
        401:
          description: Not login
        403:
          description: Permission denied
        404:
          description: The specified resource was not found
  /forum/{forumId}/alert/group:
    post:
      tags:
        - ConclusionAlert
      summary: upsert alert rules in the forum
      operationId: upsertForumConclusionAlertGroup
      parameters:
        - name: forumId
          in: path
          required: true
          schema:
            type: integer
            format: int32
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConclusionAlertGroupDetail"
      responses:
        200:
          description: Success
        400:
          description: Invalid parameters
        401:
          description: Not login
        403:
          description: Permission denied
        409:
          description: Conflict
  /forum/{forumId}/alert/rule/{ruleType}:
    post:
      tags:
        - ConclusionAlert
      summary: upsert alert rules in the forum
      operationId: upsertForumConclusionAlertRule
      parameters:
        - name: forumId
          in: path
          required: true
          schema:
            type: integer
            format: int32
        - name: ruleType
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/AlertRuleType"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConclusionAlertRuleDetail"
      responses:
        200:
          description: Success
        400:
          description: Invalid parameters
        401:
          description: Not login
        403:
          description: Permission denied
        409:
          description: Conflict
  /forum/{forumId}/alert/members:
    get:
      tags:
        - ConclusionAlert
      summary: Search for the members by their name
      operationId: searchMemberByName
      parameters:
        - name: forumId
          in: path
          required: true
          schema:
            type: integer
            format: int32
        - name: q
          in: query
          description: query name
          required: true
          schema:
            type: string
        - $ref: "#/components/parameters/LimitParam"
        - name: withGroup
          in: query
          required: true
          schema:
            type: boolean
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ConclusionAlertMember"
        400:
          description: Invalid parameters
  /topic/conclusion/alert:
    post:
      tags:
        - ConclusionAlert
      summary: alert all unconcluded topics
      operationId: alertAllUnconcludedTopics
      responses:
        200:
          description: Success
        401:
          description: Not login
        403:
          description: Permission denied
...